Math: change return type of 'math_sqrt' to i32
While it is true that the result would fit in u16, there are benefits to
using i32 instead:
- no need to clear the upper 16-bits of the return register
- return type is signed instead of unsigned

+++ include/math.h
@@ -48,7 +48,7 @@ INLINE i32 math_pow(i32 base, u32 exp) {
     return result;
 }
 
-extern u16 math_sqrt(u32 val);
+extern i32 math_sqrt(u32 val);
 
 INLINE i32 math_digits(i32 val, u32 base) {
     i32 result = 0;
